Reconfigure App Workflow
The "Main" activity after login is the MenuActivity. Currently, it only
has a couple of buttons for testing, but will soon display Leagues the
current user is a part of, and will allow the user to find or create new
leagues.

The button at the bottom allows the user to get to the old manual league
creation activities.